Split public/app.ts into smaller modules (#24)
Refactor the monolithic 2889-line app.ts into 13 focused modules:

- types.ts: TypeScript interfaces (Config, Skill, Track, Uma, SkillResult)
- constants.ts: Mapping tables (STRATEGY_TO_RUNNING_STYLE, TRACK_NAME_TO_ID)
- state.ts: Global state with getter/setter functions
- skillTrigger.ts: Skill triggering logic and restriction parsing
- skillHelpers.ts: Skill name utilities and cost calculation
- configManager.ts: Configuration I/O (loadConfig, saveConfig, autoSave)
- skillsUI.ts: Skills panel rendering
- trackUI.ts: Track panel rendering
- umaUI.ts: Uma panel rendering
- resultsUI.ts: Results table rendering and totals
- api.ts: API communication with SSE parsing
- renderCallbacks.ts: Callback registry to break circular dependencies
- app.ts: Main initialization and event setup

Key design decisions:
- State module with getters/setters for shared data access
- Render callbacks pattern to break circular dependencies between UI modules
- API callback injection to avoid circular imports

All 244 tests pass. Frontend bundle size: 101KB.

Closes #24
